refactoring. added documentation
