+++ src/main.rs
@@ -1,27 +1,21 @@
 mod driver;
 mod prepare_files;
-mod request_to_website;
+mod race_results;
 mod race;
+mod bot;
 
 use std::{
     fs,
     time::Duration,
 };
 use std::time::SystemTime;
-
-use chrono::{Datelike, Local};
 use log::LevelFilter;
-use teloxide::{
-    Bot,
-    payloads::SendMessageSetters,
-    prelude::{Message, Requester},
-    types::ParseMode,
+use crate::{
+    bot::tg_bot,
+    prepare_files::f1_official_results_links,
+    race::write_races_to_json,
 };
-use teloxide::prelude::*;
-use crate::driver::Driver;
-use crate::prepare_files::f1_official_results_links;
-use crate::race::{read_races_from_json, write_races_to_json};
-use crate::request_to_website::get_results;
+
 
 #[tokio::main]
 async fn main() {
@@ -30,13 +24,17 @@ async fn main() {
     tg_bot().await;
 }
 
+
+// Function to set up logging
 fn setup_logger() {
     pretty_env_logger::formatted_builder()
         .filter_level(LevelFilter::Info)
         .init();
 }
 
+// Function to update race json-file information
 async fn update_file_info() {
+    // Check if the race_data.json file needs to be updated
     if fs::metadata("race_data.json")
         .map_or(
             SystemTime::UNIX_EPOCH,
@@ -59,59 +57,4 @@ async fn update_file_info() {
         write_races_to_json(races).await;
     }
     log::info!("Json has the freshest results in it");
-}
-
-async fn tg_bot() {
-    log::info!("Starting f1 bot...");
-
-    let bot = Bot::from_env();
-
-
-    teloxide::repl(bot, |bot: Bot, msg: Message| async move {
-        update_file_info().await;
-        let season_races = read_races_from_json("race_data.json").unwrap_or_else(|err| {
-            log::error!("Failed to read race data: {}", err);
-            Vec::new()
-        });
-
-        for race in season_races {
-            if race.date < Local::now().naive_local().date() {
-                continue;
-            }
-            {
-                loop {
-                    let drivers_list = get_results(&race.url).await.unwrap_or_else(|err| {
-                        log::error!("Failed to get results: {}", err);
-                        Vec::new()
-                    });
-
-                    if drivers_list.is_empty() {
-                        log::info!("Race don't have results yet");
-                    }
-                    else {
-                        send_update_message(&bot, msg.chat.id, drivers_list, &race.name).await.unwrap();
-                        break
-                    }
-                    tokio::time::sleep(Duration::from_secs(1200)).await;
-                }
-            }
-        }
-        bot.send_message(msg.chat.id, "Season ended 😭\nTo receive the results of next season races, please send me /results before the next f1 year. 🏁").await.unwrap();
-
-        Ok(())
-    }).await;
-}
-
-async fn send_update_message(bot: &Bot, chat_id: ChatId, drivers: Vec<Driver>, race_name: &str) -> ResponseResult<()> {
-    let mut formatted_drivers = String::new();
-
-    formatted_drivers.push_str(&format!("{race_name}\n\n"));
-    formatted_drivers.push_str("||");
-    for driver in drivers {
-        formatted_drivers.push_str(&driver.to_string());
-    }
-    formatted_drivers.push_str("||");
-
-    bot.send_message(chat_id, formatted_drivers).parse_mode(ParseMode::MarkdownV2).await?;
-    Ok(())
 }
